+++ src/gui/plugins/transform_control/TransformControl.hh
@@ -28,7 +28,6 @@ namespace gazebo
 {
   class TransformControlPrivate;
 
-  /// \brief TODO
   /// \brief Provides buttons for translation, rotation, and scale
   ///
   /// ## Configuration
@@ -45,6 +44,28 @@ namespace gazebo
 
     // Documentation inherited
     public: void LoadConfig(const tinyxml2::XMLElement *_pluginElem) override;
+
+    /// \brief Sends an event to update the internal snapping values for
+    /// translation, rotation, and scaling.
+    /// \param[in] _x The snapping distance along the world's x (red) axis
+    /// expressed in meters.
+    /// \param[in] _y The snapping distance along the world's y (green) axis
+    /// expressed in meters.
+    /// \param[in] _z The snapping distance along the world's z (blue) axis
+    /// expressed in meters.
+    /// \param[in] _roll The snapping distance along the entity's roll (red)
+    /// axis expressed in degrees.
+    /// \param[in] _pitch The snapping distance along the entity's pitch
+    /// (green) axis expressed in degrees.
+    /// \param[in] _yaw The snapping distance along the entity's yaw (blue)
+    /// axis expressed in degrees.
+    /// \param[in] _scaleX The snapping scale along the entity's x (red) axis
+    /// \param[in] _scaleY The snapping scale along the entity's y (green) axis
+    /// \param[in] _scaleZ The snapping scale along the entity's z (blue) axis
+    public slots: void OnSnapUpdate(
+        double _x, double _y, double _z,
+        double _roll, double _pitch, double _yaw,
+        double _scaleX, double _scaleY, double _scaleZ);
 
     /// \brief Callback in Qt thread when mode changes.
     /// \param[in] _mode New transform mode