+++ src/rendering/SceneManager.cc
@@ -255,10 +255,30 @@ rendering::VisualPtr SceneManager::CreateVisual(Entity _id,
         material->SetMetalness(1.0);
       }
     }
-    // TODO(anyone) set transparency)
-    // material->SetTransparency(_visual.Transparency());
+    else
+    {
+      // meshes created by mesh loader may have their own materials
+      // update/override their properties based on input sdf element values
+      auto mesh = std::dynamic_pointer_cast<rendering::Mesh>(geom);
+      for (unsigned int i = 0; i < mesh->SubMeshCount(); ++i)
+      {
+        auto submesh = mesh->SubMeshByIndex(i);
+        auto submeshMat = submesh->Material();
+        if (submeshMat)
+        {
+          double productAlpha = (1.0-_visual.Transparency()) *
+              (1.0 - submeshMat->Transparency());
+          submeshMat->SetTransparency(1 - productAlpha);
+          submeshMat->SetCastShadows(_visual.CastShadows());
+        }
+      }
+    }
+
     if (material)
     {
+      // set transparency
+      material->SetTransparency(_visual.Transparency());
+
       // cast shadows
       material->SetCastShadows(_visual.CastShadows());
 
@@ -635,17 +655,65 @@ void SceneManager::RemoveEntity(Entity _id)
 
 /////////////////////////////////////////////////
 rendering::VisualPtr SceneManager::TopLevelVisual(
+// NOLINTNEXTLINE
     rendering::VisualPtr _visual) const
 {
-  rendering::VisualPtr rootVisual =
+  auto node = this->TopLevelNode(_visual);
+  return std::dynamic_pointer_cast<rendering::Visual>(node);
+}
+
+/////////////////////////////////////////////////
+rendering::NodePtr SceneManager::TopLevelNode(
+    const rendering::NodePtr &_node) const
+{
+  rendering::NodePtr rootNode =
       this->dataPtr->scene->RootVisual();
 
-  rendering::VisualPtr visual = std::move(_visual);
-  while (visual && visual->Parent() != rootVisual)
+  rendering::NodePtr node = _node;
+  while (node && node->Parent() != rootNode)
   {
-    visual =
-      std::dynamic_pointer_cast<rendering::Visual>(visual->Parent());
+    node =
+      std::dynamic_pointer_cast<rendering::Node>(node->Parent());
   }
 
-  return visual;
+  return node;
+}
+
+/////////////////////////////////////////////////
+Entity SceneManager::EntityFromNode(const rendering::NodePtr &_node) const
+{
+  // TODO(louise) On Citadel, set entity ID into visual with SetUserData
+  auto visual = std::dynamic_pointer_cast<rendering::Visual>(_node);
+  if (visual)
+  {
+    auto found = std::find_if(std::begin(this->dataPtr->visuals),
+        std::end(this->dataPtr->visuals),
+        [&](const std::pair<Entity, rendering::VisualPtr> &_item)
+    {
+      return _item.second == visual;
+    });
+
+    if (found != this->dataPtr->visuals.end())
+    {
+      return found->first;
+    }
+  }
+
+  auto light = std::dynamic_pointer_cast<rendering::Light>(_node);
+  if (light)
+  {
+    auto found = std::find_if(std::begin(this->dataPtr->lights),
+        std::end(this->dataPtr->lights),
+        [&](const std::pair<Entity, rendering::LightPtr> &_item)
+    {
+      return _item.second == light;
+    });
+
+    if (found != this->dataPtr->lights.end())
+    {
+      return found->first;
+    }
+  }
+
+  return kNullEntity;
 }